requirements.txt
pandas
duckdb
requests
google-cloud-bigquery
db-dtypes
functions-framework
defeatbeta-api
